Use shorter seed for box key derivation
Long seed was supposed to be benefitial because it would help to avoid
leaking the signature from which box key is derived during other uses of
the master key. Long key did not add any protection against bruteforcing
because all SSH signatures reduce inputs to some hash prior to signing
anyways.

Using master key for signing other data is not recommended now.
A project-specific prefix is now added to seed before signing to avoid
honest mistakes, but a determined attacker may still reconstruct the
message and present it for signing to master key though other channels if
the master key is reused.