Do not store box private key in memory for long time
Storing private key in RAM semi-permanently was just asking for a leak,
even with our cludgy "shield". Now we store the key only for a fraction of
second it's actually needed, and clear it immediately afterwards.
HKDF is fast enough to recompute on the fly

Another good outcome is that now decrypting each message requires a call
to ssh-agent.